most recent analysis of 200 samples of the 10k studies
This is the most up-to-date analysis output from the analyze code that uses the new Mesh integration that you wanted to add and it works like a charm. We are seeing significantly better data, there are no garbage words and we are getting very accurate outputs. This looks very promising for the project.