Added JSON for updated analyze code
This JSON analyzes the same 200, but this time the default is set to unknown, which should improve accuracy. Definitely better about accurately seeing if something is definitely there or not "present" and "absent" are more specific but the context needs to be better understood by the code. Going to mess around with the context some more to improve accuracy. Output took 4.9 seconds to generate.