Update analyze_medical_reports.py
An update that adds TargetMatcher and ConTextComponent in order for the code to be more accurate. The context component is working properly and the Target Matcher is serving the purpose of confirming the context as per the term, but for some reason our status is not reflecting the context. However, that seems to be a read issue, working on that now.