Uploaded  analysis_results_200_contextfixer2.json
This version is for the first version of the analyze code that is utilizing all of the available tools - TargetMatcher and ConTextComponent - but the accuracy is still off. However, the context reader is working very nicely and the target matcher is accurately scrubbing the data to build the context really nicely. Will continue to see how we can make our status more accurately work with the context tool.