Uploaded analysis_results_200_contextfixer5.json
This is the most accurate I have gotten the output to be. There are a couple iffy statuses. Such as, in "id": 803530, there are some instances like such: 
"text": "Pyothorax",
        "canonical": "empyema, pleural",
        "start": 508,
        "end": 517,
        "status": "present",
        "source": "mesh+spacy",
        "field": "Conclusion",
        "sentence": "Pyothorax is managed ideally with culture of fluid, chest tube placement and flushing, though sometimes non-responsive cases require advanced imaging (CT of thorax) with thoracotomy.",
        "modifiers": []

where it is more likely that the condition is present but it is still relatively inconclusive. Will look at improving the accuracy on specific cases like this, but the overall is significantly more accurate. We are slowly but surely getting to that beautiful 95% or higher accuracy.