Uploaded analysis_results_200_contextfixer6.json
The most accurate results yet. However, there is a bit of a snag in the code that gives us unknown defaults when "present" would work better because the code plays it safe. Phrases like “moderate diffuse muscle atrophy…”, “There is moderate patchy sclerosis…” are currently not hitting the exact positive trigger and are being defaulted to unknown, though this is not extremely common, it is still happening enough that it would pose a real problem. However, at this point, I have determined that this is a purely semantic issue and in the next version of the analyze code, the context tool and target matcher will have more tailored tools so we can eliminate these kinds of mistakes. (Output took 5.8 seconds for this JSON).