Uploaded analysis_results_200_contextfixer8.json
This is honestly a step backwards in terms of progress, the context tool was definitely tripping up pretty badly here. This is garbage so I'm not even going to bother with editing the analyze code to reflect what produced this. Just putting this here to show that I did something. This is basically just proof of life that I am still here trucking away at making the context tool more sophisticated. This is pretty terrible, though, I cannot lie.