Uploaded analysis_results_200_contextfixer9.json
This is one of the most accurate versions so far. Fixed up the context reader a lot but we still run into the problems of odd verbiage used in the reports as well as negations that give us a little trouble because sometimes the machine itself gets confused and thinks something is unknown if it cannot 100% verify, even if it seems obvious to us because there is a negation in there. However, this is fairly solid as is. Another version is already made and I am scanning that one over to make sure it is more accurate. Stay tuned.