Uploaded analysis_results_200_contextfixer10.json
The most recent output from the analyze code with, by far the best accuracy. It can handle simple cases with almost 100% accuracy and only really trips on obnoxious negations or statements that are a little ambiguous due to the language used when writing them. As of right now, I can confidently say that the coding and the building part are basically solidified. Now, all I am doing is refining how each individual part is moving (honestly it has been almost at this point since version 6, but now it is getting really solid).