Uploaded analysis_results_200_contextfixer11.json
This is the output that, after consulting with ChatGPT and getting its expert opinion from another, more sophisticated NLP and analyzer, I was able to determine that this is, by far and away, the most accurate reading thus far. Many unknowns were accurately verified and many incorrect absent readings were corrected to unknown. This output is the closest to 95% I have gotten (roughly 92%) and with some more tweaking on the semantic tools, this project could very well be good to go by next week.